[Silver II] Title: 통계학, Time: 68 ms, Memory: 4012 KB -BaekjoonHub

+++ 백준/Silver/2108. 통계학/통계학.cc
@@ -11,7 +11,7 @@ int arr[8001]; //최빈값 체크용도의 배열
 
 int getAvg(const vector<int>& v)
 {
-    long long ret = 0;
+    int ret = 0;
     for(auto c : v)
     {
         ret += c;
